[Refactor] Move order related models in separate module
