+++ solver/src/naive_solver.rs
@@ -3,7 +3,10 @@ mod two_order_settlement;
 use self::two_order_settlement::TwoOrderSettlement;
 use crate::settlement::Settlement;
 use contracts::{GPv2Settlement, UniswapV2Router02};
-use model::{OrderCreation, OrderKind, TokenPair};
+use model::{
+    order::{OrderCreation, OrderKind},
+    TokenPair,
+};
 use primitive_types::U512;
 use std::{cmp::Ordering, collections::HashMap};
