+++ solver/src/naive_solver/two_order_settlement.rs
@@ -3,7 +3,7 @@ use crate::{
     settlement::{Interaction, Settlement, Trade},
 };
 use contracts::{GPv2Settlement, UniswapV2Router02};
-use model::{OrderCreation, OrderKind};
+use model::order::{OrderCreation, OrderKind};
 use primitive_types::{H160, U256};
 use std::collections::HashMap;
 
@@ -174,7 +174,7 @@ fn rounded_up_division(dividend: U256, divisor: U256) -> Option<U256> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use model::OrderKind;
+    use model::order::OrderKind;
     use primitive_types::H160;
 
     fn assert_respects_prices(settlement: &TwoOrderSettlement) {
